# 8275317: AArch64: Support some type conversion vectorization in SLP