Current SLP vectorizer in C2 compiler doesn't support type conversion
operations. But AArch64 has vector type conversion instructions in
both NEON and SVE.

The type conversion involves two kinds of scenarios, conversion between
the same data sizes and conversion between different data sizes. If we
want to support casts between different data sizes, we need to amend
the code part for identifying adjacent memory references and the code
part for justifying if the combination is profitable. I suppose it
would be easier to review if we split the whole task to support type
conversion into two separate patches, one for the same data sizes and
the other one for different data sizes. The goal of this patch is just
to support conversions within the same data size, including:
  int -> float
  float -> int
  long -> double
  double -> long

A typical test case:

for (int i = start; i < limit; i++) {
    b[i] = (float) a[i];
}

To implement it, the patch completed the necessary instructions and
matching rules in the backend and added implementation for SLP in
the middle end.

The percentage of performance uplift on aarch64 system:
Mode: avgt
Cnt: 15
Metric: (ns/op)

benchmark                       percentage change  [(After-Before)/Before] 
VectorLoop.convertD2L      -48.46%	                 
VectorLoop.convertF2I        -55.67%
VectorLoop.convertI2F        -55.27%
VectorLoop.convertL2D       -48.75%

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue
- [x] Change must be properly reviewed

### Issue
 * [JDK-8275317](https://bugs.openjdk.java.net/browse/JDK-8275317): AArch64: Support some type conversion vectorization in SLP


### Reviewers
 * [Tobias Hartmann](https://openjdk.java.net/census#thartmann) (@TobiHartmann - **Reviewer**)
 * [Nick Gasson](https://openjdk.java.net/census#ngasson) (@nick-arm - **Reviewer**)


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.java.net/jdk pull/6145/head:pull/6145` \
`$ git checkout pull/6145`

Update a local copy of the PR: \
`$ git checkout pull/6145` \
`$ git pull https://git.openjdk.java.net/jdk pull/6145/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 6145`

View PR using the GUI difftool: \
`$ git pr show -t 6145`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.java.net/jdk/pull/6145.diff">https://git.openjdk.java.net/jdk/pull/6145.diff</a>

</details>